More tests for theme context
