+++ test/FieldSet.test.tsx
@@ -1,10 +1,13 @@
 import * as React from "react";
 import { render } from "@testing-library/react";
 import { FieldSet, FieldSetProps } from "../src";
-import { make } from "./helpers";
+import { make, createThemeWrapper } from "./helpers";
 
 function setup(props: Partial<FieldSetProps> = {}) {
-  return render(<FieldSet legend="Example" field={make<any>("")} {...props} />);
+  return render(
+    <FieldSet legend="Example" field={make<any>("")} {...props} />,
+    { wrapper: createThemeWrapper() }
+  );
 }
 
 describe("<FieldSet />", () => {
@@ -52,6 +55,8 @@ describe("<FieldSet />", () => {
 
   it("appends `className`", () => {
     const { container } = setup({ className: "testing" });
+
+    expect(container.firstChild).toHaveClass("fieldset");
     expect(container.firstChild).toHaveClass("testing");
   });
 
@@ -60,17 +65,23 @@ describe("<FieldSet />", () => {
       legend: "Legend",
       legendClassName: "testing"
     });
+
+    expect(getByText("Legend")).toHaveClass("legend");
     expect(getByText("Legend")).toHaveClass("testing");
   });
 
   it("appends `helpClassName`", () => {
     const { getByText } = setup({ help: "Help", helpClassName: "testing" });
+
+    expect(getByText("Help")).toHaveClass("help");
     expect(getByText("Help")).toHaveClass("testing");
   });
 
   it("appends `errorClassName`", () => {
     const field = make<any>("", { error: "Error", touched: true });
     const { getByText } = setup({ field, errorClassName: "testing" });
+
+    expect(getByText("Error")).toHaveClass("error");
     expect(getByText("Error")).toHaveClass("testing");
   });
 });
